Restructure the tests into a folder
