+++ .gitignore
@@ -1,2 +1,4 @@
 .env
 venv
+__pycache__/
+**/.DS_Store
