+++ tests/__init__.py
@@ -1,0 +1,9 @@
+from .classification import ZeroShotClassificationTest
+from .counting import CountingTest
+from .documentocr import DocumentOCRTest
+from .handwritingocr import HandwritingOCRTest
+from .extractionocr import ExtractionOCRTest
+from .mathocr import MathOCRTest
+from .objectdetection import ObjectDetectionTest
+from .setofmark import SetOfMarkTest
+from .graphunderstanding import GraphUnderstandingTest
