+++ web.py
@@ -1,12 +1,10 @@
-import base64
 import datetime
 import json
 import os
-import time
-from dataclasses import dataclass
 import re
-from Levenshtein import ratio
 from statistics import mean
+import tests
+import importlib
 
 import jinja2
 import numpy as np
@@ -17,365 +15,29 @@ from autodistill.detection import CaptionOntology, DetectionBaseModel
 
 HOME = os.path.expanduser("~")
 
-# define @running decorator that prints the name of the function it is decorating
-def running(func):
-    def wrapper(*args, **kwargs):
-        print(f"Running {func.__name__} test...")
-        return func(*args, **kwargs)
-
-    return wrapper
-
-@dataclass
-class GPT4V(DetectionBaseModel):
-    ontology: CaptionOntology
-
-    def __init__(self, ontology: CaptionOntology, api_key: str):
-        self.client = OpenAI(api_key=api_key)
-        self.ontology = ontology
-        pass
-
-    def predict(
-        self,
-        input,
-        classes,
-        result_serialization: str = "Classifications",
-        prompt: str = None,
-    ) -> sv.Classifications:
-        if prompt is None:
-            prompt = f"What is in the image? Return the class of the object in the image. Here are the classes: {', '.join(classes)}. You can only return one class from that list."
-
-        payload = [
-            {
-                "role": "user",
-                "content": [
-                    {"type": "text", "text": prompt},
-                    {
-                        "type": "image_url",
-                        "image_url": {
-                            "url": f"data:image/jpeg;base64,"
-                            + base64.b64encode(open(input, "rb").read()).decode(
-                                "utf-8"
-                            ),
-                        },
-                    },
-                ],
-            }
-        ]
-
-        start_time = time.time()
-
-        response = self.client.chat.completions.create(
-            model="gpt-4-vision-preview",
-            messages=payload,
-            max_tokens=300,
-        )
-
-        inference_time = time.time() - start_time
-
-        input_tokens = response.usage.prompt_tokens
-        output_tokens = response.usage.completion_tokens
-        tokens = (input_tokens, output_tokens)
-
-        if result_serialization == "Classifications":
-            class_ids = self.ontology.prompts().index(
-                response.choices[0].message.content
-            )
-
-            return (
-                sv.Classifications(
-                    class_id=np.array([class_ids]),
-                    confidence=np.array([1]),
-                ),
-                inference_time, tokens
-            )
-        else:
-            return response.choices[0].message.content, inference_time, tokens
-
-
-@running
-def zero_shot_classification():
-    classes = ["Tesla Model 3", "Toyota Camry"]
-
-    base_model = GPT4V(
-        ontology=CaptionOntology({"Tesla Model 3": "Tesla Model 3", "Toyota Camry": "Toyota Camry"}),
-        api_key=os.environ["OPENAI_API_KEY"],
-    )
-
-    result, inference_time, tokens = base_model.predict("images/car.jpeg", classes=classes)
-
-    return (
-        # 1 maps with Tesla Model 3
-        result == sv.Classifications(class_id=np.array([1]), confidence=np.array([1])),
-        inference_time,
-        classes[result.class_id[0]],
-        tokens
-    )
-
-
-@running
-def count_fruit():
-    base_model = GPT4V(
-        ontology=CaptionOntology({"fruit": "fruit", "bowl": "bowl"}),
-        api_key=os.environ["OPENAI_API_KEY"],
-    )
-
-    result, inference_time, tokens = base_model.predict(
-        "images/fruit.jpeg",
-        classes=["fruit", "bowl"],
-        result_serialization="text",
-        prompt="Count the fruit in the image. Return a single number.",
-    )
-
-    return result == "10", inference_time, result, tokens
-
-
-@running
-def document_ocr():
-    base_model = GPT4V(
-        ontology=CaptionOntology({"none": "none"}),
-        api_key=os.environ["OPENAI_API_KEY"],
-    )
-
-    result, inference_time, tokens = base_model.predict(
-        "images/swift.png",
-        classes=[],
-        result_serialization="text",
-        prompt="Read the text in the image. Return only the text, with puncuation."
-    )
-
-    return (
-        result == "I was thinking earlier today that I have gone through, to use the lingo, eras of listening to each of Swift's Eras. Meta indeed. I started listening to Ms. Swift's music after hearing the Midnights album. A few weeks after hearing the album for the first time, I found myself playing various songs on repeat. I listened to the album in order multiple times.",
-        inference_time, 
-        result,
-        tokens
-    )
-
-
-@running
-def handwriting_ocr():
-    base_model = GPT4V(
-        ontology=CaptionOntology({"none": "none"}),
-        api_key=os.environ["OPENAI_API_KEY"],
-    )
-
-    result, inference_time, tokens = base_model.predict(
-        "images/ocr.jpeg",
-        classes=[],
-        result_serialization="text",
-        prompt="Read the text in the image. Return only the text, with puncuation."
-    )
-
-    return (
-        result == 'The words of songs on the album have been echoing in my head all week. "Fades into the grey of my day old tea."',
-        inference_time,
-        result,
-        tokens
-    )
-
-
-@running
-def extraction_ocr():
-    base_model = GPT4V(
-        ontology=CaptionOntology({"none": "none"}),
-        api_key=os.environ["OPENAI_API_KEY"],
-    )
-
-    result, inference_time, tokens = base_model.predict(
-        "images/prescription.png",
-        classes=[],
-        result_serialization="text",
-        prompt="Return a JSON array containing information about the prescription in this image. Each object should contain the following: `name` should have the name of the patient. `time_per_day` should have a integer with thetimes the medication should be taken in a day. `medication` should have the brand name of the medication. `dosage` should have a integer in mg units of each tablet. `rx_number` should have the prescription number, also marked Rx. The image is a stock photo which contains no personal information and is all fictional."
-    )
-
-    code_regex = r'```[a-zA-Z]*\n(.*?)\n```'
-    code_blocks = re.findall(code_regex,result, re.DOTALL)
-    if (len(code_blocks) == 0): 
-        return 0, inference_time, f"Failed to produce a valid JSON output: {result}", tokens
-    answer_array = json.loads(code_blocks[0])
-
-    correct_array = [
-        {
-            "name": "MARY THOMAS",
-            "time_per_day": 1,
-            "medication": "ATENOLOL",
-            "dosage": 100,
-            "rx_number": "1234567-12345"
-        }
-    ]
-
-    accuracy = ratio(str(answer_array).lower(), str(correct_array).lower())
-    return accuracy, inference_time, str(answer_array), tokens
-
-
-@running
-def math_ocr():
-    base_model = GPT4V(
-        ontology=CaptionOntology({"none": "none"}),
-        api_key=os.environ["OPENAI_API_KEY"],
-    )
-
-    result, inference_time, tokens = base_model.predict(
-        "images/math.jpeg",
-        classes=[],
-        result_serialization="text",
-        prompt="Produce a JSON array with a LaTeX string of each equation in the image."
-    )
-
-    code_regex = r'```[a-zA-Z]*\n(.*?)\n```'
-    code_blocks = re.findall(code_regex,result, re.DOTALL)
-    if (len(code_blocks) == 0): 
-        return 0, inference_time, f"Failed to produce a valid JSON output: {result}", tokens
-    answer_array = json.loads(code_blocks[0])
-    answer_equation = answer_array[0].replace(" ", "")
-
-    correct_equation = "3x^2-6x+2"
-
-    accuracy = ratio(str(answer_equation).lower(), str(correct_equation).lower())
-    return accuracy, inference_time, str(answer_equation), tokens
-
-
-@running
-def object_detection():
-    base_model = GPT4V(
-        ontology=CaptionOntology({"none": "none"}),
-        api_key=os.environ["OPENAI_API_KEY"],
-    )
-
-    result, inference_time, tokens = base_model.predict(
-        "images/fruit.jpeg",
-        classes=[],
-        result_serialization="text",
-        prompt="If there are banana in this image, return a JSON object with `x`, `y`, `width` and `height` properties of the banana. All values should be normalized between 0-1 and x&y should be the center point.",
-    )
-
-    code_regex = r'```[a-zA-Z]*\n(.*?)\n```'
-    code_blocks = re.findall(code_regex, result, re.DOTALL)
-    if (len(code_blocks) == 0): 
-        return 0, inference_time, f"Failed to produce a valid JSON output: {result}", tokens
-    answer = json.loads(code_blocks[0])
-
-    correct = {'x': 0.465, 'y': 0.42, 'width': 0.37, 'height': 0.38}
-
-    r1 = answer
-    r2 = correct
-    xi_min = max(r1['x'] - r1['width'] / 2, r2['x'] - r2['width'] / 2)
-    yi_min = max(r1['y'] - r1['height'] / 2, r2['y'] - r2['height'] / 2)
-    xi_max = min(r1['x'] + r1['width'] / 2, r2['x'] + r2['width'] / 2)
-    yi_max = min(r1['y'] + r1['height'] / 2, r2['y'] + r2['height'] / 2)
-
-    inter_area = max(0, xi_max - xi_min) * max(0, yi_max - yi_min)
-    union_area = r1['width'] * r1['height'] + r2['width'] * r2['height'] - inter_area
-
-    iou = inter_area / union_area if union_area else 0
-
-    return iou, inference_time, str(answer), tokens
-
-@running
-def set_of_mark():
-    base_model = GPT4V(
-        ontology=CaptionOntology({"none": "none"}),
-        api_key=os.environ["OPENAI_API_KEY"],
-    )
-
-    result, inference_time, tokens = base_model.predict(
-        "images/fruits_som.png",
-        classes=[],
-        result_serialization="text",
-        prompt="Find all the fruits in this image and return a JSON array of all the applicable numbers.",
-    )
-
-    code_regex = r'```[a-zA-Z]*\n(.*?)\n```'
-    code_blocks = re.findall(code_regex, result, re.DOTALL)
-    if (len(code_blocks) == 0): 
-        return 0, inference_time, f"Failed to produce a valid JSON output: {result}", tokens
-    answer = json.loads(code_blocks[0])
-
-    correct = [35,40,26,2,13,17,29,21,10,42,8,43,0,11,7,4,12,27,37,39,22,15,25]
-
-    score = 0
-    for guess in answer:
-        if guess in correct: score += 1
-
-    accuracy = score/len(correct)
-
-    return accuracy, inference_time, str(answer), tokens
-
-
-def graph_understanding():
-    base_model = GPT4V(
-        ontology=CaptionOntology({"none": "none"}),
-        api_key=os.environ["OPENAI_API_KEY"],
-    )
-
-    result, inference_time, tokens = base_model.predict(
-        "images/graph.png",
-        classes=[],
-        result_serialization="text",
-        prompt="State positions of points A through D in a JSON with properties A-D, each having a object with properties for integers matching the respective point: `quantity` and `price`.",
-    )
-
-    code_regex = r'```[a-zA-Z]*\n(.*?)\n```'
-    code_blocks = re.findall(code_regex, result, re.DOTALL)
-    if (len(code_blocks) == 0): 
-        return 0, inference_time, f"Failed to produce a valid JSON output: {result}", tokens
-    answer = json.loads(code_blocks[0])
-
-    correct = {
-      "A": {
-        "quantity": 20,
-        "price": 10
-      },
-      "B": {
-        "quantity": 26,
-        "price": 20
-      },
-      "C": {
-        "quantity": 30,
-        "price": 30
-      },
-      "D": {
-        "quantity": 34,
-        "price": 40
-      }
-    }
-
-    total_scores = 0
-    count = 0
-
-    for letter in 'ABCD':
-        if letter in correct and letter in answer:
-            quantity_diff = abs(correct[letter]['quantity'] - answer[letter]['quantity'])
-            quantity_score = max(0, 1 - (quantity_diff / 25))
-
-            price_diff = abs(correct[letter]['price'] - answer[letter]['price'])
-            price_score = max(0, 1 - (price_diff / 25))
-
-            total_scores += (quantity_score + price_score)
-            count += 2
-
-    print(total_scores / count)
-    score = total_scores / count
-
-    return score, response_time, str(result), tokens
-
-
-tests = [
-    "zero_shot_classification",
-    "count_fruit",
-    "document_ocr",
-    "handwriting_ocr",
-    "extraction_ocr",
-    "math_ocr",
-    "object_detection",
-    "set_of_mark",
-    "graph_understanding"
+test_list = [
+    "ZeroShotClassificationTest",
+    "CountingTest",
+    "DocumentOCRTest",
+    "HandwritingOCRTest",
+    "ExtractionOCRTest",
+    "MathOCRTest",
+    "ObjectDetectionTest",
+    "SetOfMarkTest",
+    "GraphUnderstandingTest"
 ]
 
-current_results = {}
-for i in tests:
-    test = globals()[i]
+test_ids = []
 
-    test_result = test()
+current_results = {}
+for i in test_list:
+    test_info = getattr(importlib.import_module(f"tests"),i)
+    print(f"Running {test_info.name} test...")
+
+    test_id = test_info.id
+    test_ids.append(test_id)
+
+    test_result = test_info.test()
     score, response_time, result, tokens = test_result
 
     input_token_price = 0.01/1000
@@ -383,13 +45,13 @@ for i in tests:
     price = (input_token_price * tokens[0]) + (output_token_price * tokens[1])
     score = (1 if score is True else (0 if score is False else score))
 
-    current_results[i] = {}
-    current_results[i]["score"] = score
-    current_results[i]["success"] = score == 1
-    current_results[i]["price"] = price
-    current_results[i]["pass_fail"] = "Pass" if score == 1 else "Fail"
-    current_results[i]["response_time"] = response_time
-    current_results[i]["result"] = result
+    current_results[test_id] = {}
+    current_results[test_id]["score"] = score
+    current_results[test_id]["success"] = score == 1
+    current_results[test_id]["price"] = price
+    current_results[test_id]["pass_fail"] = "Pass" if score == 1 else "Fail"
+    current_results[test_id]["response_time"] = response_time
+    current_results[test_id]["result"] = result
 
 print("current_results", current_results)
 
@@ -403,11 +65,22 @@ today = datetime.datetime.now().strftime("%Y-%m-%d")
 with open(f"results/{today}.json", "w+") as file:
     json.dump(current_results, file, indent=4)
 
-historical_results = {}
-for i in tests:
-    historical_results[i] = {}
-    historical_results[i]["scores"] = []
-    historical_results[i]["response_times"] = []
+results = {}
+
+for index, test_id in enumerate(test_ids):
+    results[test_id] = {}
+
+    test_info = getattr(importlib.import_module(f"tests"),test_list[index])
+    results[test_id]["name"] = test_info.name
+    results[test_id]["question"] = test_info.question
+    results[test_id]["prompt"] = test_info.prompt
+    results[test_id]["image"] = test_info.image
+    results[test_id]["method"] = test_info.method
+
+for i in test_ids:
+    results[i]["history"] = {}
+    results[i]["history"]["scores"] = []
+    results[i]["history"]["response_times"] = []
 
 for file in os.listdir("results"):
     if os.path.isdir(f"results/{file}"): continue
@@ -416,39 +89,27 @@ for file in os.listdir("results"):
 
         for key, value in data.items():
             print(key, value)
-            if historical_results.get(key) is None: continue
-            historical_results[key]["scores"].append(value["score"])
-            historical_results[key]["response_times"].append(value["response_time"])
-            historical_results[key]["days"] = len(historical_results[key]["scores"])
+            if results.get(key) is None: continue
+            results[key]["history"]["scores"].append(value["score"])
+            results[key]["history"]["response_times"].append(value["response_time"])
+            results[key]["history"]["days"] = len(results[key]["history"]["scores"])
 
-
-print("historical_results", historical_results)
-
-historical_averages = {}
-for i in tests:
-    historical_averages[i] = {}
-    historical_averages[i]["scores"] = mean(historical_results[i]["scores"])
-    historical_averages[i]["response_times"] = mean(historical_results[i]["response_times"])
-    historical_averages[i]["success_percent"] = round(historical_averages[i]["scores"]*100,2)
-
-print("historical_averages", historical_averages)
+for i in test_ids:
+    results[i]["average"] = {}
+    results[i]["average"]["score"] = mean(results[i]["history"]["scores"])
+    results[i]["average"]["response_time"] = mean(results[i]["history"]["response_times"])
+    results[i]["average"]["success_percent"] = round(results[i]["average"]["score"]*100,2)
 
 response_times = []
-for i in historical_averages:
-    response_times.append(historical_averages[i]["response_times"])
+for i in test_ids:
+    response_times.append(results[i]["average"]["response_time"])
 
-average_response_time = round(mean(response_times),2)
+average_response_time = round(mean(response_times), 2)
 day_count = len(response_times)
 
-results = {
-    'current': current_results,
-    'past': historical_results,
-    'averages': historical_averages,
-    'avg_time': average_response_time,
-    'days': day_count
-}
-
-print(json.dumps(result, indent=4))
+print("- - - - -")
+print(json.dumps(results, indent=4))
+print("- - - - -")
 
 template = jinja2.Template(open("template.html").read())
